% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{by_isolation_forest_plot}
\alias{by_isolation_forest_plot}
\title{Plot per-sample anomaly scores from the isolation forest}
\usage{
by_isolation_forest_plot(output_anomaly_df, title = "")
}
\arguments{
\item{output_anomaly_df}{Tibble returned in `anomaly_df` from
[by_isolation_forest()].}

\item{title}{Plot title.}
}
\value{
A `plotly` figure (printed when invoked at top level).
}
\description{
Plot per-sample anomaly scores from the isolation forest
}
\keyword{internal}
