% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{revert_log_transform}
\alias{revert_log_transform}
\title{Inverse of log2-transform: raise intensities to the power of 2}
\usage{
revert_log_transform(df)
}
\arguments{
\item{df}{Long-format intensity data frame.}
}
\value{
`df` with `Intensity = 2^Intensity`.
}
\description{
Inverse of log2-transform: raise intensities to the power of 2
}
\keyword{internal}
