% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spqrp-package.R
\docType{package}
\name{spqrp-package}
\alias{spqrp}
\alias{spqrp-package}
\title{spqrp: Sample Provenance Quality Resolver in Proteomics}
\description{
Detects sample-provenance inconsistencies in MS-based plasma-proteome
cohorts via pairwise distance, threshold-based classification, iterative
clustering, and a pairwise random-forest classifier for protein importance
ranking. Native R port of the Python package of the same name.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fhradilak/spqrp_r}
  \item Report bugs at \url{https://github.com/fhradilak/spqrp_r/issues}
}

}
\author{
\strong{Maintainer}: Franziska Hradilak \email{Franziska.Hradilak@student.hpi.uni-potsdam.de}

Authors:
\itemize{
  \item Franziska Hradilak \email{Franziska.Hradilak@student.hpi.uni-potsdam.de}
}

}
\keyword{internal}
