% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spqrp_example_path}
\alias{spqrp_example_path}
\title{Filesystem path to a bundled example CSV}
\usage{
spqrp_example_path(which = c("input_cohort_df", "protein_ranking"))
}
\arguments{
\item{which}{One of `"input_cohort_df"`, `"protein_ranking"`.}
}
\value{
Absolute character path inside `inst/extdata/`.
}
\description{
Filesystem path to a bundled example CSV
}
\examples{
spqrp_example_path("input_cohort_df")
}
